% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{train_test_split}
\alias{train_test_split}
\title{A Train/Test Split Function}
\usage{
train_test_split(x, y, validation_split)
}
\arguments{
\item{x}{(data frame): A data frame of predictors.}

\item{y}{(list/vector): A list or vector of responses.}

\item{validation_split}{(numeric): A number between 0 and 1
that represents the proportion of the data to be used for testing.}
}
\value{
out (list): A list containing all of the below components...

$train_x (data frame): A data frame of predictors for the
training set.

$train_y (list/vector): A list or vector of responses for
the training set.

$test_x (data frame): A data frame of predictors for the testing set.


}
\description{
This function splits the data into training and testing sets.
}
